requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0
django-ckeditor==5.9.0
django-extensions==3.1.5
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.2.0
django-user-visit==0.4
gunicorn==20.0.4
psycopg2==2.7.7
psycopg2-binary==2.9.2
pygraphviz==1.7
pytz==2019.3
sqlparse==0.3.1
ua-parser==0.10.0
user-agents==2.2.0
whitenoise==5.0.1
